==> cc3m_experiments/00_filter_empty_files.ipynb <==
from pathlib import Path
import os
from tqdm.auto import tqdm

# filter all the empty files from the conceptual captions dataset. Empty files are created when the download of the images failed.
path = Path('./data/conceptual_captions/cc_data/val')

image_files = [*path.glob("**/*.png"), *path.glob("**/*.jpg"), *path.glob("**/*.jpeg"), *path.glob("**/*.bmp")]

empty_files = []
for pth in tqdm(image_files):
    if os.stat(pth).st_size == 0:
        empty_files.append(pth)

for pth in empty_files:
    os.remove(pth)
